fix: mobile need login modal
